requirements.txt
numpy==1.23.3
pycryptodome==3.15.0